Moved Pinecone Pydantic types to their own file
